Everything working for now last version

+++ app/controllers/products_controller.rb
@@ -1,6 +1,6 @@
 class ProductsController < ApplicationController
   before_action :set_product, only: [:show, :update, :destroy]
-  skip_before_action :get_current_user, only: [:index, :show]
+  skip_before_action :get_current_user, only: [:index, :show, :search]
 
   # GET /products
   def index
@@ -25,6 +25,52 @@ class ProductsController < ApplicationController
       else render json: trans.errors, status: :unprocessable_entity end
     else render json: product.errors, status: :unprocessable_entity end
   end
+  
+  def search
+    users = nil
+    products = nil
+    products_price_range = nil
+    if !params[:input].nil?
+      params[:input] = params[:input].downcase if params[:input].is_a?(String)
+      users = User.search(params[:input])
+      products = if params[:input].is_a?(String) then
+        if params[:input].include?("auction") or params[:input].include?("auctions") then
+          Product.where(is_auction:true).order("created_at DESC")
+        else
+          Product.search(params[:input])
+        end
+      else
+        Product.search(params[:input])
+      end
+    end
+    if !params[:price_range].nil? and !products.nil?
+      price_range = params[:price_range].split("-")
+      products_price_range = products.where("price BETWEEN ? AND ?", price_range[0],  price_range[1]).order("created_at DESC")
+    end
+    if !params[:auction].nil? and !products.nil?
+      if params[:auction]
+        products = products.where(is_auction:true).order("created_at DESC")
+      else
+        products = products.where(is_auction:false).order("created_at DESC")
+      end
+    end
+    if !params[:category].nil? and !products.nil?
+        products = products.where(category:params[:category]).order("created_at DESC")
+    end
+    if !users.nil? and !products.nil?
+      render json: {
+        users: ActiveModel::Serializer::CollectionSerializer.new(users, serializer: UserSerializer),  
+        products: ActiveModel::Serializer::CollectionSerializer.new(products, serializer: ProductSerializer),
+        products_price_range: if !products_price_range.nil? then
+          ActiveModel::Serializer::CollectionSerializer.new(products_price_range, serializer: ProductSerializer)
+        else 
+          'No results for price range'
+        end
+        }, status: :ok
+    else
+      render json: {empty: 'no results'}, status: :unprocessable_entity
+    end
+  end
 
   # PATCH/PUT /products/1
   def update
@@ -40,12 +86,14 @@ class ProductsController < ApplicationController
 
   # DELETE /products/1
   def destroy
-    if @product.user_id == @current_user.id
-      render_ok @product.destroy
-    elsif is_current_user_admin.nil?
-      render_ok @product.destroy
-    end
-  end
+    if product_does_not_have_purchases?
+       if @product.user_id == @current_user.id
+         render_ok @product.destroy
+       elsif is_current_user_admin.nil?
+         render_ok @product.destroy
+       end
+     end
+   end
 
   private
     # Use callbacks to share common setup or constraints between actions.
